Support implementation of component classes.
Support class implementations extending React.Component.

The type of props is extracted similarly to pure export declarations.

Support ignoring some statements present in implementations, but not essential to extract the type of components.

Support inlining of prop types.

Extend treatment of type tables with a global table of all the toplevel type definitions.

Add tests for components defined with and without inlining the type of props.